move chevron link to correct folder
